+++ src/libraries.lisp
@@ -32,13 +32,15 @@
       (if lib
         (progn (setf *coreclr-location* lib)
                (pushnew (uiop:pathname-directory-pathname lib)
-                        *foreign-library-directories*))
+                        *foreign-library-directories*
+                        :test #'equalp))
         (error "Unable to find CoreCLR")))
     (let ((lib (find-interop t)))
       (if lib
         (progn (setf *interop-location* lib)
                (pushnew (uiop:pathname-directory-pathname lib)
-                        *foreign-library-directories*))
+                        *foreign-library-directories*
+                        :test #'equalp))
         (error "Unable to find BikeInterop.dll")))
     (values))
   (uiop:register-image-restore-hook 'init-coreclr-search-location))
@@ -47,19 +49,108 @@
   (define-foreign-library coreclr
     (t #.+coreclr-library-file+))
 
-  (use-foreign-library coreclr))
+  (unless (foreign-library-loaded-p 'coreclr)
+    (use-foreign-library coreclr)))
+
+(eval-when (:compile-toplevel :load-toplevel :execute)
+  (defun %get-related-app-path (name)
+    (cl-ppcre:regex-replace
+     "(?i)Microsoft\\.NetCore\\.App"
+     (native-path (uiop:pathname-directory-pathname *coreclr-location*))
+     name)))
+
+#+coreclr-windows
+(progn
+  (eval-when (:compile-toplevel :load-toplevel :execute)
+    (defvar *desktop-sdk-dir* nil)
+    (defun initialize-wpfgfx-search-location ()
+      (let* ((desktop-sdk-dir (pathname
+                               (%get-related-app-path "Microsoft.WindowsDesktop.App"))))
+        (if (uiop:probe-file*
+             (uiop:make-pathname* :name "wpfgfx_cor3.dll"
+                                  :defaults desktop-sdk-dir))
+          (pushnew (setf *desktop-sdk-dir* desktop-sdk-dir)
+                   *foreign-library-directories*
+                   :test #'equalp)
+          (setf *desktop-sdk-dir* nil))))
+    (uiop:register-image-restore-hook 'initialize-wpfgfx-search-location))
+
+  (eval-when (:compile-toplevel :load-toplevel :execute)
+    ;; These libs should be pre-loaded, otherwise WPF would be unable
+    ;;   to find them, for who knows what reasons
+
+    (define-foreign-library vcruntime
+      (t "vcruntime140_cor3.dll"))
+    (define-foreign-library d3dcompiler
+      (t "D3DCompiler_47_cor3.dll"))
+    (define-foreign-library wpfgfx
+      (t "wpfgfx_cor3.dll"))
+    (defun load-wpfgfx ()
+      (when *desktop-sdk-dir*
+        (unless (foreign-library-loaded-p 'vcruntime)
+          (use-foreign-library vcruntime))
+        (unless (foreign-library-loaded-p 'd3dcompiler)
+          (use-foreign-library d3dcompiler))
+        (unless (foreign-library-loaded-p 'wpfgfx)
+          (use-foreign-library wpfgfx))))
+    (uiop:register-image-restore-hook 'load-wpfgfx)))
+
+(defun %get-app-paths ()
+  (let ((directories
+          (delete-duplicates
+           (mapcar #'native-path
+                   (remove nil
+                           (list
+                            (uiop:pathname-directory-pathname *coreclr-location*)
+                            #+windows
+                            *desktop-sdk-dir*
+                            (uiop:pathname-directory-pathname *interop-location*)
+                            (uiop:get-pathname-defaults)
+                            (uiop:lisp-implementation-directory)
+                            (uiop:pathname-directory-pathname (get-exe-path)))))
+           :test #'equalp)))
+    (format nil (uiop:strcat "~{~a~^" (uiop:inter-directory-separator) "~}")
+            (nreverse directories))))
 
 (defun get-trusted-platform-assemblies ()
   "Retrieves a list of pathnames of trusted platform assemblies"
-  (uiop:directory* (uiop:make-pathname*
-                    :name uiop:*wild*
-                    :type "dll"
-                    :defaults *coreclr-location*)))
+  (let ((defaults (uiop:directory* (uiop:make-pathname*
+                                    :name uiop:*wild*
+                                    :type "dll"
+                                    :defaults *coreclr-location*))))
+    ;; On Windows, we want to default to Desktop SDK, should
+    ;;  it be installed, to be able to use WinForms/WPF.
+    ;; How's the addition of the desktop SDK dir to APP/NI
+    ;;  paths is not enough?
+    ;; Well, the thing is, those folders contain different
+    ;;  WindowsBase.dll assembly, amongst others, and should we
+    ;;  only include the coreclr directory for TPA assemblies, then WindowsBase
+    ;;  would be not the WindowsBase the PresentationCore.dll etc are looking for.
+    ;; So hence we want to default to the one that is present inside Desktop SDK dir.
+    ;; It wont hurt apps that won't use WPF/WinForms, since Desktop SDK assemblies are
+    ;;  simply extended ones of the defaults.
+    #+coreclr-windows
+    (if *desktop-sdk-dir*
+      (loop :with desktop-asms = (uiop:directory*
+                                  (uiop:make-pathname*
+                                   :name uiop:*wild*
+                                   :type "dll"
+                                   :defaults *desktop-sdk-dir*))
+            :for asm :in defaults
+            :for replacement = (find-if (lambda (name)
+                                          (string-equal (pathname-name asm)
+                                                        name))
+                                        desktop-asms
+                                        :key #'pathname-name)
+            :if replacement :collect replacement
+              :else :collect asm)
+      defaults)
+    #-coreclr-windows
+    defaults))
 
 (defun %get-tpa-string ()
   (format nil (uiop:strcat "~{~a~^" (uiop:inter-directory-separator) "~}~a~a")
-          (mapcar #'native-path
-                  (get-trusted-platform-assemblies))
+          (mapcar #'native-path (get-trusted-platform-assemblies))
           (uiop:inter-directory-separator)
           (native-path *interop-location*)))
 
@@ -70,14 +161,5 @@
         :when (and (uiop:string-prefix-p "System." name)
                    (not (uiop:string-suffix-p name ".Native")))
           :collect name))
-
-(defun %get-app-paths ()
-  (format nil (uiop:strcat "~{~a~^" (uiop:inter-directory-separator) "~}")
-          (mapcar #'native-path
-                  (list (uiop:get-pathname-defaults)
-                        (uiop:lisp-implementation-directory)
-                        (uiop:pathname-directory-pathname (get-exe-path))
-                        (uiop:pathname-directory-pathname *interop-location*)
-                        (uiop:pathname-directory-pathname *coreclr-location*)))))
 
 ;;; vim: ft=lisp et